fix(execute): splits execute and opcodes
